+++ CHANGELOG.md
@@ -5,6 +5,14 @@ file is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and
 this project adheres to [Semantic
 Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Added
+
+- `AgentInfo` to represent the agent's information during the handshake process.
+- `server::handshake` implements the application-level handshake process for the
+  server after a QUIC connection is established.
+
 ## [0.1.2] - 2024-03-25
 
 ### Added
@@ -44,6 +52,7 @@ Versioning](https://semver.org/spec/v2.0.0.html).
 - `client::handshake` implements the application-level handshake process for the
   client after a QUIC connection is established.
 
+[Unreleased]: https://github.com/petabi/review-protocol/compare/0.1.2...main
 [0.1.2]: https://github.com/petabi/review-protocol/compare/0.1.1...0.1.2
 [0.1.1]: https://github.com/petabi/review-protocol/compare/0.1.0...0.1.1
 [0.1.0]: https://github.com/petabi/review-protocol/tree/0.1.0